backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
celery[redis]==5.4.0
redis==5.2.1
python-multipart==0.0.20
websockets==14.1
aiofiles==24.1.0
Pillow==11.1.0
pydub==0.25.1
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
greenlet==3.1.1
bcrypt==4.2.1
PyJWT==2.10.1
stripe==12.1.0
eval_type_backport==0.2.2
